requirements.txt
streamlit
opencv-python
face-recognition
numpy
pandas
mysql-connector-python
python-dotenv
pillow 